Extend expire time for queued jobs
This commit extends the expiry time for queued jobs from 1 minute to 60
minutes. The reason being that prior to pipelines, we only ever expected
queued jobs to be picked up immediately without any downstream jobs.
This no longer makes sense given that Pipelines can queue a LOT of jobs
at once and it is reasonable that they might wait a while before
executing.

A future todo here is to set a 60 second time out if a runner is running
and has accepted a job. There's no reason we should wait in that
instance.