internal/server: Support reserved pipeline step "deploy"
This commit introduces the deploy reserved built-in step for Pipelines.
It supports a Release bool, but intentionally does not implement release
yet as that will be coming in a follow-up commit/pull request to this
one.

It also introduces a change in the deploy operation in that if there was
no artifact set on the jobs Deploy operation, it looks up the latest
pushed artifact to use that. This change is important because pipelines
queues its jobs ahead of time all at once, which means if at any point a
Build step runs prior, we couldn't look up the artifact without it
existing. So the Deploy step will not set the artifact, making it nil,
and when the Deploy goes to run it will look up the latest.